moved HDF5 routines into actual folders and proper places
